+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/avm2/parser/script/ActionScript3Parser.java
@@ -150,6 +150,8 @@ public class ActionScript3Parser {
 //    private final AbcIndexing otherABCs;
     //private static final List<ABC> playerABCs = new ArrayList<>();
     private static AbcIndexing playerGlobalAbcIndex;
+    
+    private static AbcIndexing airGlobalAbcIndex;
 
     private long uniqId() {
         uniqLast++;
@@ -2591,10 +2593,10 @@ public class ActionScript3Parser {
         addScriptFromTree(allOpenedNamespaces, traits, classPos);
     }
 
-    public ActionScript3Parser(ABC abc, List<ABC> otherAbcs) throws IOException, InterruptedException {
+    public ActionScript3Parser(ABC abc, List<ABC> otherAbcs, boolean air) throws IOException, InterruptedException {
         initPlayer();
 
-        abcIndex = new AbcIndexing(playerGlobalAbcIndex);
+        abcIndex = new AbcIndexing(air ? airGlobalAbcIndex : playerGlobalAbcIndex);
         for (ABC a : otherAbcs) {
             abcIndex.addAbc(a);
         }
@@ -2612,12 +2614,20 @@ public class ActionScript3Parser {
             SWF swf = new SWF(swc.getSWF("library.swf"), true);
             playerGlobalAbcIndex = new AbcIndexing(swf);
         }
+        if (airGlobalAbcIndex == null) {
+            if (Configuration.getAirSWC() == null) {
+                return;
+            }
+            SWC swc = new SWC(new FileInputStream(Configuration.getAirSWC()));
+            SWF swf = new SWF(swc.getSWF("library.swf"), true);
+            airGlobalAbcIndex = new AbcIndexing(swf);
+        }
     }
 
-    public static void compile(String src, ABC abc, List<ABC> otherABCs, String fileName, int classPos, int scriptIndex) throws AVM2ParseException, IOException, InterruptedException, CompilationException {
+    public static void compile(String src, ABC abc, List<ABC> otherABCs, String fileName, int classPos, int scriptIndex, boolean air) throws AVM2ParseException, IOException, InterruptedException, CompilationException {
         //List<ABC> parABCs = new ArrayList<>();
         initPlayer();
-        ActionScript3Parser parser = new ActionScript3Parser(abc, otherABCs);
+        ActionScript3Parser parser = new ActionScript3Parser(abc, otherABCs, air);
         boolean success = false;
         ABC originalAbc = ((ABCContainerTag) ((Tag) abc.parentTag).cloneTag()).getABC();
         try {
@@ -2638,12 +2648,12 @@ public class ActionScript3Parser {
         }
     }
 
-    public static void compile(SWF swf, String src, String dst, int classPos, int scriptIndex) {
+    public static void compile(SWF swf, String src, String dst, int classPos, int scriptIndex, boolean air) {
         System.err.println("WARNING: AS3 compiler is not finished yet. This is only used for debuggging!");
         try {
             initPlayer();
             ABC abc = new ABC(null);
-            ActionScript3Parser parser = new ActionScript3Parser(abc, new ArrayList<>());
+            ActionScript3Parser parser = new ActionScript3Parser(abc, new ArrayList<>(), air);
             parser.addScript(new String(Helper.readFile(src), Utf8Helper.charset), src, classPos, scriptIndex);
             try (OutputStream fos = new BufferedOutputStream(new FileOutputStream(new File(dst)))) {
                 abc.saveToStream(fos);