+++ src/com/jpexs/decompiler/flash/gui/Main.java
@@ -202,7 +202,21 @@ public class Main {
     public static CancellableWorker importWorker = null;
     public static CancellableWorker deobfuscatePCodeWorker = null;
     public static CancellableWorker swfPrepareWorker = null;
+    
+    public static final int LIBRARY_AIR = 0;
+    public static final int LIBRARY_FLASH = 1;
 
+    
+    public static boolean isSwfAir(SWF swf) {
+        SwfSpecificCustomConfiguration conf = Configuration.getSwfSpecificCustomConfiguration(swf.getShortPathTitle());
+        if (conf != null) {
+            String libraryAsStr = conf.getCustomData(CustomConfigurationKeys.KEY_LIBRARY, "" + LIBRARY_FLASH);
+            int libraryAsInt = Integer.parseInt(libraryAsStr);
+            return libraryAsInt == LIBRARY_AIR;
+        }
+        return false;
+    }
+    
     //This method makes file watcher to shut up during our own file saving
     public static void startSaving(File savedFile) {
         savedFiles.add(savedFile);