+++ src/com/jpexs/decompiler/flash/gui/tagtree/TagTreeContextMenu.java
@@ -1614,7 +1614,7 @@ public class TagTreeContextMenu extends JPopupMenu {
                     }
                 }
                 try {
-                    ActionScript3Parser parser = new ActionScript3Parser(doAbc.getABC(), abcs);
+                    ActionScript3Parser parser = new ActionScript3Parser(doAbc.getABC(), abcs, false);
 
                     DottedChain dc = new DottedChain(pkgParts);
                     String script = "package " + dc.toPrintableString(true) + " {"