+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - [#1875] Remove no longer accessed items from cache after certain amount of time
 - [#1280] AS3 Direct editation of traits with the same name
 - [#1743] GFX - Adding DefineExternalImage2 and DefineSubImage tags
+- [#1822], [#1803] AS3 direct editation - optional using AIR (airglobal.swc) to compile
 
 ### Fixed
 - [#1869] Replace references now replaces all references, not just PlaceObject
@@ -2600,6 +2601,8 @@ All notable changes to this project will be documented in this file.
 [#1875]: https://www.free-decompiler.com/flash/issues/1875
 [#1280]: https://www.free-decompiler.com/flash/issues/1280
 [#1743]: https://www.free-decompiler.com/flash/issues/1743
+[#1822]: https://www.free-decompiler.com/flash/issues/1822
+[#1803]: https://www.free-decompiler.com/flash/issues/1803
 [#1869]: https://www.free-decompiler.com/flash/issues/1869
 [#1872]: https://www.free-decompiler.com/flash/issues/1872
 [#1692]: https://www.free-decompiler.com/flash/issues/1692